【AC】[M-SOLUTIONS プロコンオープン]A - Sum of Interior Angles

+++ m-solutions2019/m_solutions2019_a.cpp
@@ -46,7 +46,7 @@ int main() {
     int n;
     cin >> n;
 
-    double ans = 180.0 * (n - 2);
-    printf("%.20f\n", ans);
+    int ans = 180 * (n - 2);
+    printf("%d\n", ans);
 
 }
